+++ nowcasting_dataloader/data_sources/topographic/topographic_model.py
@@ -66,9 +66,9 @@ class TopographicML(DataSourceOutputML):
                 batch_size,
                 image_size_pixels,
                 image_size_pixels,
-            ),
-            topo_x_coords=np.sort(np.random.randn(batch_size, image_size_pixels)),
-            topo_y_coords=np.sort(np.random.randn(batch_size, image_size_pixels))[:, ::-1].copy(),
+            ).astype(np.float32),
+            topo_x_coords=np.sort(np.random.randn(batch_size, image_size_pixels).astype(np.float32)),
+            topo_y_coords=np.sort(np.random.randn(batch_size, image_size_pixels).astype(np.float32))[:, ::-1].copy(),
             # copy is needed as torch doesnt not support negative strides
         )
